pyproject.toml
[project]
name = "vc-notice-bot"
version = "0.1.0"
description = "VCの入退室を報告するbot"
authors = [
    { name = "isksss", email = "104404522+isksss@users.noreply.github.com" }
]
dependencies = [
    "discord-py>=2.3.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/vc_notice_bot"]
